requirements.txt
python-dotenv
azure-storage-blob
azure-storage-file-datalake
pyodbc
tabulate